Feat(Java): Move unreferenced SAML Beans out of AuthenticationHandler
